% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alleles_recessive_test.R
\name{all_enrichment}
\alias{all_enrichment}
\title{test enrichment of inherited biallelic LoF, compound heterozygous LoF/Func or biallelic Func}
\usage{
all_enrichment(freq, count, cohort_n, autozygosity = 0)
}
\arguments{
\item{freq}{list of cumulative frequencies of variation in a population for
rare LoF variants, and rare functional variants.}

\item{count}{number of probands with inherited variants in the gene.}

\item{cohort_n}{number of probands in population.}

\item{autozygosity}{rate of autozygosity in the cohort being investigated.}
}
\value{
P-value from testing for biallelic LoF, Lof/Func and biallelic Func variants.
}
\description{
test enrichment of inherited biallelic LoF, compound heterozygous LoF/Func or biallelic Func
}
\examples{
freq = list(lof=0.001, functional=0.1)
count = 3
lof_func_enrichment(freq, count, 1000)

}

