% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/alleles_recessive_test.R
\name{analyse_inherited_enrichment}
\alias{analyse_inherited_enrichment}
\title{test for enrichment of inherited variants in the DDD and ExAC datasets}
\usage{
analyse_inherited_enrichment(hgnc, chrom, biallelic_lof, biallelic_func,
  lof_func, probands = NULL, cohort_n = 3072, check_last_base = FALSE,
  autozygous_rate = 0)
}
\arguments{
\item{hgnc}{HGNC symbol for a gene.}

\item{chrom}{chromosome that the gene is on.}

\item{biallelic_lof}{number of probands with inherited biallelic LoF variants
in the gene.}

\item{biallelic_func}{number of probands with inherited biallelic functional
variants in the gene.}

\item{lof_func}{number of probands with inherited Lof/Func variants in the gene.}

\item{probands}{vector of probands who have inherited recessive variants in
the gene, or NULL.}

\item{cohort_n}{number of probands in population.}

\item{check_last_base}{whether to correct missense or synonymous G alleles at
the last base of exons to a LoF consequence.}

\item{autozygous_rate}{rate of autozygosity within the gene in the probands.}
}
\value{
a list of P values from tests using the DDD population, the ExAC
    population, under LoF and functional tests.
}
\description{
test for enrichment of inherited variants in the DDD and ExAC datasets
}

