% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suspected_syndromes_test.R
\name{calculate_gene_similarity}
\alias{calculate_gene_similarity}
\title{calculate the similarity of suspected syndrome terms within a group of probands}
\usage{
calculate_gene_similarity(probands, syndrome_rates)
}
\arguments{
\item{probands}{list of probands, with their suspected syndromes as vectors}

\item{syndrome_rates}{dataframe of syndrome terms and the rates at which they
are used within the DDD probands.}
}
\value{
summed similarity across the probands
}
\description{
calculate the similarity of suspected syndrome terms within a group of probands
}

