% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/suspected_syndromes_test.R
\name{calculate_no_match}
\alias{calculate_no_match}
\title{calculate the proportion of samples with no matches}
\usage{
calculate_no_match(syndromes)
}
\arguments{
\item{syndromes}{list of syndrome terms per proband}
}
\value{
proportion of sample pairs lacking a syndrome term match
}
\description{
calculate the proportion of samples with no matches
}

