% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/suspected_syndromes_test.R
\name{calculate_proband_similarity}
\alias{calculate_proband_similarity}
\title{calculate the similarity of suspected syndromes between two probands}
\usage{
calculate_proband_similarity(syndrome_rates, proband_1, proband_2)
}
\arguments{
\item{syndrome_rates}{dataframe of syndrome terms and the rates at which they
are used within the DDD probands.}

\item{proband_1}{vector of syndrome terms for a proband}

\item{proband_2}{vector of syndrome terms for a proband}
}
\value{
the similarity of the suspected syndromes for two syndromes, as rate
    between 0 and 1.
}
\description{
calculate the similarity of suspected syndromes between two probands
}

