% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suspected_syndromes_test.R
\name{calculate_syndrome_rates}
\alias{calculate_syndrome_rates}
\title{calculate the rate at which each syndrome term is used within the population}
\usage{
calculate_syndrome_rates(syndromes)
}
\arguments{
\item{syndromes}{list of syndrome terms per proband}
}
\value{
dataframe of unique terms per row, with the rate at which
}
\description{
calculate the rate at which each syndrome term is used within the population
}

