% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_autozygosity.R
\name{check_sample_autozygosity}
\alias{check_sample_autozygosity}
\title{identify whether a proband is autozygous within a genome region}
\usage{
check_sample_autozygosity(bcf_path, chrom, start_pos, end_pos, proband)
}
\arguments{
\item{bcf_path}{path to a BCF for the genome region}

\item{chrom}{chromosome that we wish to analyse}

\item{start_pos}{nucleotide position that starts the genome region}

\item{end_pos}{nucleotide position that end the genome region}

\item{proband}{ID for the proband to analyse}
}
\value{
True/False for whether the sample is autozygous within the region
}
\description{
identify whether a proband is autozygous within a genome region
}

