% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculate_autozygosity.R
\name{check_sample_autozygosity_genome_wide}
\alias{check_sample_autozygosity_genome_wide}
\title{identify the genome regions where a proband has runs-of-homozygosity}
\usage{
check_sample_autozygosity_genome_wide(bcf_path, proband)
}
\arguments{
\item{bcf_path}{path to a BCF for the full genome across all probands}

\item{proband}{ID for the proband to analyse}
}
\value{
dataframe of proband ID, chromosome, start and end coordinates.
}
\description{
identify the genome regions where a proband has runs-of-homozygosity
}

