% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/alleles_ddd.R
\name{convert_genotypes}
\alias{convert_genotypes}
\title{reformats the genotypes dataset}
\usage{
convert_genotypes(vars, vep, probands)
}
\arguments{
\item{vars}{seqminer::readVCFToListByRange output, list of values, which
includes a genotype matrix as "GT", and sample IDs as sampleId.}

\item{vep}{dataframe of VEP consequences for the vars entry}

\item{probands}{vector of proband IDs, or NA, or a zero-length vector.}
}
\value{
dataframe of variants
}
\description{
reformats the genotypes dataset
}

