% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_frequency.R
\name{cumulative_frequency}
\alias{cumulative_frequency}
\title{calculate cumulative allele frequency from a vector of allele frequencies}
\usage{
cumulative_frequency(probs)
}
\arguments{
\item{probs}{vector of frequencies}
}
\value{
summed cumulative frequency.
}
\description{
Rather than simply summing the individual allele frequencies to get a
cumulative frequency, we need to take the chance of having two variants on
one chromosome into account. The chance of observing two variants on the same
chromosome is their frequencies multiplied together (e.g. p1 * p2), so the
probability of seeing them on independent chromosomes is p1 * (1 - p2).
This can be extended beyond two variants, to simply account for the
probability of the preceeding variants.
}
\examples{

freqs = runif(100, min=1e-7, max=0.01)
cumulative_frequency(freqs)
}

