% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alleles_recessive_test.R
\name{enrichment_single_population}
\alias{enrichment_single_population}
\title{test for enrichment of inherited variants}
\usage{
enrichment_single_population(freq, counts, cohort_n, autozygosity = 0)
}
\arguments{
\item{freq}{list of cumulative frequencies of variation in a population for
rare LoF variants, and rare functional variants.}

\item{counts}{list of counts of biallelically inherited genotypes in a gene for:
biallelic_lof: inherited Lof/LoF variants.
lof_func: ninherited Lof/Func variants.
biallelic_func: inherited func/func variants.}

\item{cohort_n}{number of probands in population.}

\item{autozygosity}{rate of autozygosity in the cohort being investigated.}
}
\value{
a list of P values from tests, under LoF and functional tests.
}
\description{
test for enrichment of inherited variants
}
\examples{
freq = list(lof=0.001, functional=0.1)
counts = list("biallelic_lof"=1, 'biallelic_func'=4, 'lof_func'=2)
enrichment_single_population(freq, counts, 1000)

}

