% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculate_autozygosity.R
\name{extract_region}
\alias{extract_region}
\title{create a BCF for a small genome region, for a limited number of probands}
\usage{
extract_region(chrom, start_pos, end_pos, vcf_dir)
}
\arguments{
\item{chrom}{chromosome that we wish to analyse}

\item{start_pos}{nucleotide position that starts the genome region}

\item{end_pos}{nucleotide position that end the genome region}

\item{vcf_dir}{directory containing multi-sample VCFs per-chromosome for the
population of interest.}
}
\value{
path to the newly created BCF
}
\description{
create a BCF for a small genome region, for a limited number of probands
}

