% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_CADD.R
\name{filter_exac_variants_by_cadd}
\alias{filter_exac_variants_by_cadd}
\title{filter ExAC variants by CADD score}
\usage{
filter_exac_variants_by_cadd(vars,
  path = "/lustre/scratch115/projects/ddd/users/hcm/DDD/CADD_resources/ExAC_r0.3.tsv.gz",
  threshold = 21, exclude = FALSE,
  tempdir = "/lustre/scratch115/projects/ddd/users/hcm/DDD/tempdir_for_gene_vcfs_for_annotation/",
  notLoFs = TRUE)
}
\arguments{
\item{vars}{data frame or list of data frames with ExAC variants, output from get_exac_variants_for_gene}

\item{path}{path to vcf.gz file of CADD scores, default = /lustre/scratch115/projects/ddd/users/hcm/DDD/CADD_resources/ExAC_r0.3.tsv.gz}

\item{threshold}{CADD score cutoff for variants, default = 21}

\item{exclude}{boolean for whether to exclude variants without a CADD score, default=FALSE}

\item{tempdir}{path to temporary diectory for writing VCF files; default = /lustre/scratch115/projects/ddd/users/hcm/DDD/tempdir_for_gene_vcfs_for_annotation/}

\item{notLoFs}{boolean for whether to keep all LoFs regardless of CADD score, default = TRUE}
}
\value{
data frame or list of data frame  with ExAC variants, filtered by CADD
}
\description{
filter ExAC variants by CADD score
}

