% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/alleles_ddd.R
\name{find_file_pos}
\alias{find_file_pos}
\title{find where the required data lies within the file}
\usage{
find_file_pos(path, start_pos)
}
\arguments{
\item{path}{path to the datafile}

\item{start_pos}{start position of the variant}
}
\value{
the approximate byte offset in the file (to within a few hundred
   bytes) where the required data starts
}
\description{
Seek in a file to the section containing the data that we want. We start
with a reasonable guess about where in the file the data is located, but
we seek up or downstream of that (depending on whether the data at that
position is before or after the data that we want), then bisect to pin
down where to start loading data from.
}

