% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gencode}
\alias{gencode}
\title{table of genecode gene coordinates}
\format{A data frame with two columns variables: \code{chr}, \code{start},
    \code{stop}, \code{gene_id}, \code{transcript_id}, \code{gene},
    \code{status}, \code{level} and \code{gene_type}.}
\source{
url(http://www.gencodegenes.org/releases/19.html)
}
\usage{
gencode
}
\description{
Tom Fitzgerald provided this table, where the positions were derived from a
GENCODE data release.
}
\keyword{datasets}

