% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_autozygosity.R
\name{get_autozygous_rate}
\alias{get_autozygous_rate}
\title{identify whether a proband is autozygous within a genome region}
\usage{
get_autozygous_rate(hgnc, chrom, vcf_dir = DDD_VCFS_DIR)
}
\arguments{
\item{hgnc}{HGNC symbol for the gene of interest}

\item{chrom}{chromosome that we wish to analyse}

\item{vcf_dir}{directory containing multi-sample VCFs per-chromosome for the
population of interest.}
}
\value{
True/False for whether the sample is autozygous within the region
}
\description{
identify whether a proband is autozygous within a genome region
}

