% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/alleles_recessive_test.R
\name{get_count_combinations}
\alias{get_count_combinations}
\title{Get all the combinations of spreading the probands across populations.}
\usage{
get_count_combinations(populations, biallelic_lof, biallelic_func, lof_func)
}
\arguments{
\item{populations}{a vector of population names to be tested}

\item{biallelic_lof}{number of probands with inherited biallelic LoF variants
in the gene.}

\item{biallelic_func}{number of probands with inherited biallelic functional
variants in the gene.}

\item{lof_func}{number of probands with inherited Lof/Func variants in the gene.}
}
\value{
a list of count dataframes for each functional type
}
\description{
Get all the combinations of spreading the probands across populations.
}

