% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/alleles_ddd.R
\name{get_ddd_cohort}
\alias{get_ddd_cohort}
\title{get the details of the DDD cohort}
\usage{
get_ddd_cohort(parents = TRUE, unaffected = TRUE)
}
\arguments{
\item{parents}{boolean for whether to restrict to parents only.}

\item{unaffected}{boolean for whether to restrict to unaffected only.}
}
\value{
vector of paths to VCFs
}
\description{
get the details of the DDD cohort
}

