% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alleles_ddd.R
\name{get_ddd_vep_annotations}
\alias{get_ddd_vep_annotations}
\title{gets the DDD VEP annotations for variants}
\usage{
get_ddd_vep_annotations(chrom, start, end)
}
\arguments{
\item{chrom}{chromosome as character string}

\item{start}{start of gene range}

\item{end}{end of gene range}
}
\value{
dataframe of VEP annotations
}
\description{
gets the DDD VEP annotations for variants
}

