% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alleles_exac.R
\name{get_exac_variants_for_gene}
\alias{get_exac_variants_for_gene}
\title{get and parse ExAC variants within a gene}
\usage{
get_exac_variants_for_gene(hgnc, chrom, start = NULL, end = NULL,
  fileName = NULL, check_last_base = FALSE, remove_benign = FALSE)
}
\arguments{
\item{hgnc}{HGNC symbol for the gene that we want data for eg "ARID1B"}

\item{chrom}{chromosome string e.g. "6"}

\item{start}{start position of region to be investigated (if checking for
gene region defined by chromosome coordinates rather than a HGNC-based
gene region).}

\item{end}{end position of region to be investigated (if checking for
gene region defined by chromosome coordinates rather than a HGNC-based
gene region).}

\item{fileName}{path to tabis-indexed ExAC VCF file.}

\item{check_last_base}{boolean for whether to check if last base non-lofs can
be LoF.}

\item{remove_benign}{boolean for whether to exclude missense_variants with
polyphen tolerated predictions.}
}
\value{
data frame of variants in gene
}
\description{
get and parse ExAC variants within a gene
}

