% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tweak_lofs.R
\name{get_exon_ends}
\alias{get_exon_ends}
\title{gets all exon end positions for an HGNC symbol}
\usage{
get_exon_ends(hgnc_symbol)
}
\arguments{
\item{hgnc_symbol}{HGNC symbol for gene of interest.}
}
\value{
list of vectors of exon end chromosome positions.
}
\description{
gets all exon end positions for an HGNC symbol
}

