% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gene_coordinates.R
\name{get_gene_coordinates}
\alias{get_gene_coordinates}
\title{get the genomic coordinates for a gene from an internal gencode dataframe}
\usage{
get_gene_coordinates(hgnc, chrom)
}
\arguments{
\item{hgnc}{HGNC symbol for the gene that we want data for eg "ARID1B"}

\item{chrom}{chromosome string e.g. "6"}
}
\value{
data frame of genomic cooridantes
}
\description{
get the genomic coordinates for a gene from an internal gencode dataframe
}

