% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tweak_lofs.R
\name{get_gene_ids_for_hgnc}
\alias{get_gene_ids_for_hgnc}
\title{gets the gene IDs for HGNC symbols}
\usage{
get_gene_ids_for_hgnc(base_url, hgnc_symbol)
}
\arguments{
\item{base_url}{url for REST server.}

\item{hgnc_symbol}{HGNC symbol for gene of interest.}
}
\value{
vector of ensembl gene IDs.
}
\description{
gets the gene IDs for HGNC symbols
}

