% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alleles_ddd.R
\name{get_hemizygous}
\alias{get_hemizygous}
\title{reformat to hemizygous chrX genotypes in a single column}
\usage{
get_hemizygous(column)
}
\arguments{
\item{column}{column of genotypes}
}
\value{
column of variants, where male allosomal chrX genotypes have been
    converted to hemizygous genotypes (e.g. "0/0" -> "0").
}
\description{
The column has previously been vetted as for a male, on chromosome X, not
within a pseudoautosomal region.
}

