% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alleles_ddd.R
\name{get_lines_from_vep_vcf}
\alias{get_lines_from_vep_vcf}
\title{loads DDD VEP annotations lines for variants}
\usage{
get_lines_from_vep_vcf(chrom, start, end)
}
\arguments{
\item{chrom}{chromosome as character string}

\item{start}{start of gene range}

\item{end}{end of gene range}
}
\value{
dataframe of VEP annotations
}
\description{
loads DDD VEP annotations lines for variants
}

