% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/alleles_exac.R
\name{get_list_of_population_frequencies}
\alias{get_list_of_population_frequencies}
\title{create a list of allele frequency dataframes for each population.}
\usage{
get_list_of_population_frequencies(vars, populations)
}
\arguments{
\item{vars}{data frame of variants in gene, with allele counts and totals for
several ExAC sub-groups.}

\item{populations}{list of populations (e.g. list(EAS="East Asian",
FIN="Finnish")). Perhaps this should be a simple vector of names, but
this way I have more information about what the populations are.}
}
\value{
list of dataframes, one for each population. The dataframes contain
        allele count, allele number and VEP consequence strings for each
        allele in the input dataframe. The list is named for each of the
        populations.
}
\description{
create a list of allele frequency dataframes for each population.
}

