% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse_vep.R
\name{get_most_severe_consequence}
\alias{get_most_severe_consequence}
\title{get the most severe VEP consequence form a list of consequences}
\usage{
get_most_severe_consequence(consequences)
}
\arguments{
\item{consequences}{vector of VEP consequence strings. Can be an empty vector}
}
\value{
most severe VEP consequence string, or NA if the consequences vector
    is empty.
}
\description{
get the most severe VEP consequence form a list of consequences
}

