% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_polyphen.R
\name{get_polyphen}
\alias{get_polyphen}
\title{get a polyphen prediction for a variant}
\usage{
get_polyphen(chrom, start, end, alt)
}
\arguments{
\item{chrom}{chromosome that the variant is on (e.g. "8")}

\item{start}{start nucleotide position of the variant}

\item{end}{end nucleotide position of the variant}

\item{alt}{alternate allele code (e.g. "C")}
}
\value{
the polyphen prediction for a variant.
}
\description{
get a polyphen prediction for a variant
}

