% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tweak_lofs.R
\name{get_transcript_ids_for_ensembl_gene_ids}
\alias{get_transcript_ids_for_ensembl_gene_ids}
\title{gets the transcript IDs for ensembl gene IDs}
\usage{
get_transcript_ids_for_ensembl_gene_ids(base_url, gene_ids, hgnc)
}
\arguments{
\item{base_url}{url for REST server.}

\item{gene_ids}{vector of gene IDs.}

\item{hgnc}{HGNC symbol for gene of interest.}
}
\value{
vector of transcript IDs.
}
\description{
gets the transcript IDs for ensembl gene IDs
}

