% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculate_autozygosity.R
\name{get_undiagnosed_sanger_ids}
\alias{get_undiagnosed_sanger_ids}
\title{identify DDD probands to analyse}
\usage{

  get_undiagnosed_sanger_ids(vcf_path = "/lustre/scratch113/projects/ddd/users/jm33/ddd_4k.bcftools.bcf")
}
\arguments{
\item{vcf_path}{path to multi-sample VCFs for the population of interest.}
}
\description{
identify DDD probands to analyse
}

