% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alleles_recessive_test.R
\name{lof_func_enrichment}
\alias{lof_func_enrichment}
\title{test enrichment of inherited biallelic LoF and compound heterozygous LoF/Func}
\usage{
lof_func_enrichment(freq, count, cohort_n, autozygous_rate = 0)
}
\arguments{
\item{freq}{list of cumulative frequencies of variation in a population for
rare LoF variants, and rare functional variants.}

\item{count}{number of probands with inherited variants in the gene.}

\item{cohort_n}{number of probands in population.}

\item{autozygous_rate}{rate of autozygosity in the cohort being investigated.}
}
\value{
P-value from testing for biallelic LoF and Lof/Func variants.
}
\description{
test enrichment of inherited biallelic LoF and compound heterozygous LoF/Func
}

