% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_function.R
\name{mytest}
\alias{mytest}
\title{obtain cumulative allele frequency of rare lof and functional variants}
\usage{
mytest()
}
\description{
obtain cumulative allele frequency of rare lof and functional variants
}

