% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suspected_syndromes_test.R
\name{open_suspected_syndromes}
\alias{open_suspected_syndromes}
\title{open suspected syndromes}
\usage{
open_suspected_syndromes(path)
}
\arguments{
\item{path}{path to curated list of suspected syndromes per proband}
}
\value{
list of probands, where each proband has a vector of suspected
    syndromes.
}
\description{
open suspected syndromes
}

