% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_vep.R
\name{parse_vep_output}
\alias{parse_vep_output}
\title{parses VEP output}
\usage{
parse_vep_output(variant, hgnc, remove_benign = TRUE)
}
\arguments{
\item{variant}{single row dataframe containing allele column and VEP prediction
column.}

\item{hgnc}{HGNC symbol for the gene that we are currently investigating.}

\item{remove_benign}{boolean for whether to exclude missense_variants with
polyphen tolerated predictions.}
}
\value{
vep consequence string for the most severe consequence for the gene
    of interest. If none of the predicted consequences lie within the gene of
   interest, return "NA". If the variant has multiple alleles, return a
   comma-separated list of VEP consequences for the alleles.
}
\description{
parses VEP output
}

