% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/suspected_syndromes_test.R
\name{permute_probands}
\alias{permute_probands}
\title{permutes the probands between genes}
\usage{
permute_probands(probands)
}
\arguments{
\item{probands}{list of probands per gene}
}
\value{
permuted list of probands per gene
}
\description{
permutes the probands between genes
}

