% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recessiveStats-package.R
\docType{package}
\name{recessiveStats}
\alias{recessiveStats}
\alias{recessiveStats-package}
\title{recessiveStats: examines enrichment of recesive variants within genes.}
\description{
recessiveStats: examines enrichment of recesive variants within genes.
}

