% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/alleles_ddd.R
\name{reformat_chrX_genotypes}
\alias{reformat_chrX_genotypes}
\title{reformat chrX genotypes for males to hemizygous}
\usage{
reformat_chrX_genotypes(vars, geno, ddd_parents)
}
\arguments{
\item{vars}{seqminer::readVCFToListByRange output, list of values, which
includes a genotype matrix as "GT", and sample IDs as sampleId.}

\item{geno}{dataframe of biallelic genotypes, one column per individual}

\item{ddd_parents}{information about DDD individuals, including sex}
}
\value{
dataframe of variants, where male allosomal chrX genotypes have been
    converted to hemizygous genotypes (e.g. "0/0" -> "0").
}
\description{
Female chrX genotypes are untouched, as are genotypes on autosomal
chromosomes. Males with heterozygous chrX genotypes are left unchanged, as we
cannot be sure which genotype is correct. Possibly they should be removed. We
also do not alter genotypes within pseduo-autosomal regions.
}

