% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alleles_ddd.R
\name{remove_easy_nonfunctional}
\alias{remove_easy_nonfunctional}
\title{quickly strips out the variants which are obviously nonfunctional}
\usage{
remove_easy_nonfunctional(vars, vep)
}
\arguments{
\item{vars}{seqminer::readVCFToListByRange output, list of values, which
includes a genotype matrix as "GT", and sample IDs as sampleId.}

\item{vep}{dataframe of VEP consequences for the vars entry}
}
\value{
dataframe of variants
}
\description{
quickly strips out the variants which are obviously nonfunctional
}

