% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tidy_variants.R
\name{remove_nonfunctional_variants}
\alias{remove_nonfunctional_variants}
\title{exclude nonfunctional variants, or nonfunctional alleles of variants}
\usage{
remove_nonfunctional_variants(vars)
}
\arguments{
\item{vars}{dataframe of variants, with consequence, alt allele, and allele
count columns.}
}
\value{
dataframe of variants with nonfunctional variants/alleles trimmed
}
\description{
exclude nonfunctional variants, or nonfunctional alleles of variants
}

