% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/set_exac_path.R
\name{set_exac_path}
\alias{set_exac_path}
\title{sets the path to the ExAC dataset files}
\usage{
set_exac_path(path)
}
\arguments{
\item{path}{path to the ExAC VCF.}
}
\description{
sets the path to the ExAC dataset files
}

