% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tidy_variants.R
\name{standardise_multiple_alt_variants}
\alias{standardise_multiple_alt_variants}
\title{tidy up a variant dataset, so that multiple alt variants are in standardised rows}
\usage{
standardise_multiple_alt_variants(vars, include_hgnc = FALSE)
}
\arguments{
\item{vars}{dataframe of variants, with consequence, alt allele, and allele
count columns.}

\item{include_hgnc}{whether the dataframe also includes a HGNC column that
needs to be standardised.}
}
\value{
dataframe of variants with mutlple alt alleles spread across separate
    rows.
}
\description{
Make a new dataframe for the multiple alt variants, where each row is
for a single alt allele, so alleles at a single site span multiple rows.
}

