% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/alleles_recessive_test.R
\name{sum_combo_tests}
\alias{sum_combo_tests}
\title{get a p-value that sums across different population possibilities}
\usage{
sum_combo_tests(exac, cohort_n, populations, combos, p_name)
}
\arguments{
\item{exac}{list of frequency estimates for each ExAC population.}

\item{cohort_n}{list of number of probands in each population.}

\item{populations}{a vector of population names to be tested.}

\item{combos}{a dataframe of the possible count combinations for a functional
type.}

\item{p_name}{the name of the p-value to extract (e.g. "biallelic_lof_p").}
}
\value{
a p-value from testing for
}
\description{
get a p-value that sums across different population possibilities
}

