% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SYNDROMES}
\alias{SYNDROMES}
\title{table of syndromes, with regexes to match in the probands recorded syndromes}
\format{A data frame with two columns variables: \code{regex} and \code{name}}
\usage{
SYNDROMES
}
\description{
table of syndromes, with regexes to match in the probands recorded syndromes
}
\keyword{datasets}

