% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/alleles_ddd.R
\name{tally_alleles}
\alias{tally_alleles}
\title{count alleles for a variant}
\usage{
tally_alleles(variant)
}
\arguments{
\item{variant}{single row of data frame of VCF genotypes}
}
\value{
dataframe of variants
}
\description{
count alleles for a variant
}

