% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alleles_recessive_test.R
\name{analyse_inherited_enrichment}
\alias{analyse_inherited_enrichment}
\title{test for enrichment of inherited variants}
\usage{
analyse_inherited_enrichment(counts, variants, cohort_n = 3072,
  autozygosity = 0)
}
\arguments{
\item{counts}{list of counts of biallelically inherited genotypes in a gene for:
biallelic_lof: inherited Lof/LoF variants.
lof_func: ninherited Lof/Func variants.
biallelic_func: inherited func/func variants.}

\item{variants}{dataframe of variants in gene}

\item{cohort_n}{number of probands in population.}

\item{autozygosity}{rate of autozygosity within the gene in the probands.}
}
\value{
a list of P values from LoF and functional tests.
}
\description{
test for enrichment of inherited variants
}

