% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alleles_recessive_test.R
\name{biallelic_lof_enrichment}
\alias{biallelic_lof_enrichment}
\title{test for enrichment of biallelic LoF inherited variants}
\usage{
biallelic_lof_enrichment(freq, count, cohort_n, autozygosity = 0)
}
\arguments{
\item{freq}{list of cumulative frequencies of variation in a population for
rare LoF variants, and rare functional variants.}

\item{count}{number of probands with inherited variants in the gene.}

\item{cohort_n}{number of probands in population.}

\item{autozygosity}{rate of autozygosity in the cohort being investigated.}
}
\value{
P-value from testing for biallelic LoF variants.
}
\description{
test for enrichment of biallelic LoF inherited variants
}
\examples{
freq = list(lof=0.001, functional=0.1)
count = 3
biallelic_lof_enrichment(freq, count, 1000)

}

