% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/alleles_recessive_test.R
\name{biallelic_lof_enrichment}
\alias{biallelic_lof_enrichment}
\title{test for enrichment of biallelic LoF inherited variants}
\usage{
biallelic_lof_enrichment(freq, count, cohort_n)
}
\arguments{
\item{freq}{list of cumulative frequencies of variation in a population for
rare LoF variants, and rare functional variants.}

\item{count}{number of probands with inherited variants in the gene.}

\item{cohort_n}{number of probands in population.}
}
\value{
P-value from testing for biallelic LoF variants.
}
\description{
test for enrichment of biallelic LoF inherited variants
}

