% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alleles_recessive_test.R
\name{get_count_combinations}
\alias{get_count_combinations}
\title{Get all the combinations of spreading the probands across populations.}
\usage{
get_count_combinations(populations, count)
}
\arguments{
\item{populations}{a vector of population names to be tested}

\item{count}{number of probands with inherited variants in the gene.}
}
\value{
a list of count dataframes for each functional type
}
\description{
Get all the combinations of spreading the probands across populations.
}

