% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cumulative_frequency.R
\name{get_cumulative_frequencies}
\alias{get_cumulative_frequencies}
\title{obtain cumulative allele frequency of rare lof and functional variants}
\usage{
get_cumulative_frequencies(vars)
}
\arguments{
\item{vars}{dataframe of variants (one row per allele), which includes the
number of times that allele was observed within the population, as
well as the total number of alleles in the population.}
}
\value{
a list of loss of function cumulative frequency, and functional
        cumulative frequency.
}
\description{
obtain cumulative allele frequency of rare lof and functional variants
}
\examples{
vars = read.table(header = TRUE, text = "
     AC  AN    CQ
     1   1000  missense_variant
     1   1000  stop_gained
     1   1000  stop_lost
     1   1000  synonymous_variant")
get_cumulative_frequencies(vars)
}

