% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/alleles_ddd.R
\name{get_ddd_variants_for_gene}
\alias{get_ddd_variants_for_gene}
\title{loads the variants for a given gene from source VCFs}
\usage{
get_ddd_variants_for_gene(hgnc, chrom, probands = NULL,
  check_last_base = FALSE)
}
\arguments{
\item{hgnc}{hgnc symbol as character string}

\item{chrom}{chromosome as character string}

\item{probands}{vector of IDs for probands for the gene, or NULL}

\item{check_last_base}{boolean for whether to check if last base non-lofs can
be LoF.}
}
\value{
dataframe of variants
}
\description{
loads the variants for a given gene from source VCFs
}

