% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculate_autozygosity.R
\name{get_undiagnosed_sanger_ids}
\alias{get_undiagnosed_sanger_ids}
\title{identify DDD probands to analyse}
\usage{
get_undiagnosed_sanger_ids(chrom, vcf_dir = DDD_VCFS_DIR)
}
\arguments{
\item{chrom}{chromosome of interest (this will only be important if we have
different sets of probands for each chromosome file, which shouldn't
be the case, but we might as well be sure.)}

\item{vcf_dir}{directory containing multi-sample VCFs per-chromosome for the
population of interest.}
}
\description{
identify DDD probands to analyse
}
\examples{
get_undiagnosed_sanger_ids("14")
}

