% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alleles_recessive_test.R
\name{sum_combo_tests}
\alias{sum_combo_tests}
\title{get a p-value that sums across different population possibilities}
\usage{
sum_combo_tests(exac, cohort_n, combos, enrich_function, autozygous_rate = 0)
}
\arguments{
\item{exac}{list of frequency estimates for each ExAC population.}

\item{cohort_n}{list of number of probands in each population.}

\item{combos}{a dataframe of the possible count combinations for a functional
type.}

\item{enrich_function}{function to test enrichment.}

\item{autozygous_rate}{rate of autozygosity in the cohort being investigated.}
}
\value{
a p-value from testing for
}
\description{
get a p-value that sums across different population possibilities
}

