% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alleles_recessive_test.R
\name{sum_combo_tests}
\alias{sum_combo_tests}
\title{get a p-value that sums across different population possibilities}
\usage{
sum_combo_tests(freqs, cohort_n, combos, enrich_function, autozygosity = 0)
}
\arguments{
\item{freqs}{list of frequency estimates for each ExAC population.}

\item{cohort_n}{list of number of probands in each population.}

\item{combos}{a dataframe of the possible count combinations for a functional
type.}

\item{enrich_function}{function to test enrichment.}

\item{autozygosity}{rate of autozygosity in the cohort being investigated.}
}
\value{
a p-value from testing for
}
\description{
get a p-value that sums across different population possibilities
}
\examples{
freqs = list("A"=list("lof"=0.1, "functional"=0.1),
    "B"=list("lof"=0.1, "functional"=0.1))
cohort_n = list("A"=100, "B"=100)
combos = get_count_combinations(names(cohort_n), count=1)
sum_combo_tests(freqs, cohort_n, combos, biallelic_lof_enrichment)

}

