% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/alleles_recessive_test.R
\name{test_enrichment_across_multiple_populations}
\alias{test_enrichment_across_multiple_populations}
\title{test for enrichment of inherited variants in multiple populations}
\usage{
test_enrichment_across_multiple_populations(exac, biallelic_lof, biallelic_func,
  lof_func, cohort_n)
}
\arguments{
\item{exac}{list of frequency estimates for each ExAC population}

\item{biallelic_lof}{number of probands with inherited biallelic LoF variants in the gene.}

\item{biallelic_func}{number of probands with inherited biallelic functional variants in the gene.}

\item{lof_func}{number of probands with inherited Lof/Func variants in the gene.}

\item{cohort_n}{list of number of probands in each population.}
}
\value{
a list of P values from tests across the populations, under LoF and
        functional tests.
}
\description{
test for enrichment of inherited variants in multiple populations
}

